backend/requirements.txt
fastapi
uvicorn
flask
pydantic
httpx
python-dotenv
yt-dlp
google-generativeai
